move nondefault hlgroups to a different folder
usage : extended_integrations = { "bufferline", "alpha" } in M.ui table, this was necessary. Because in the futuer i will be adding more hlgroup support for another non default plugins and it will be costly to load all the hlgroups by default. So i let the user mention it in his chadrc regarding custom hlgroups he wants